pyproject.toml
[tool.poetry]
name = "reform"
version = "0.4"
description = "Reform is a tool kit developed to expand and simplify creating terraform infrastructures with the AWS provider."
authors = ["Aaron Collins <Aaron.Collinsa+reform@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
invoke = ">=1.6.0"
boto3 = ">=1.27.0"
pycryptodome = "^3.10.1"
pyee = "^8.1.0"
Jinja2 = ">=3.0.1"
python-hcl2 = "^2.0.3"
pyyaml = "^6.0.1"


[tool.poetry.dev-dependencies]
black = "^20.8b1"
nose = "^1.3.7"

[tool.poetry.scripts]
reform = 'reform:program.run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
